Rename secret to keyring
